pyproject.toml
[project]
name = "echi-listeningtests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gdown>=5.2.0",
    "hydra-core>=1.3.2",
    "jiwer>=4.0.0",
    "numpy<2.0.0",
    "pandas>=2.3.2",
    "ppgs>=0.0.9",
    "pre-commit>=4.3.0",
    "pysepm",
    "pystoi>=0.4.1",
    "soundfile>=0.13.1",
    "soxr>=1.0.0",
    "streamlit>=1.50.0",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
pysepm = { git = "https://github.com/schmiph2/pysepm" }
